[migrations] fixed migrations to skip non existing tables

+++ db/migrate/20180124091123_add_field_template_name_and_schema.rb
@@ -1,6 +1,9 @@
 class AddFieldTemplateNameAndSchema < ActiveRecord::Migration[5.0]
   def up
     DataCycleCore.content_tables.each do |table_name|
+
+      next unless @connection.table_exists?(table_name)
+
       add_column table_name.to_sym, :template_name, :string
       add_column table_name.to_sym, :schema, :jsonb
       add_column (table_name.singularize + '_histories').to_sym, :template_name, :string
@@ -10,6 +13,9 @@ class AddFieldTemplateNameAndSchema < ActiveRecord::Migration[5.0]
 
   def down
     DataCycleCore.content_tables.each do |table_name|
+
+      next unless @connection.table_exists?(table_name)
+
       remove_column table_name.to_sym, :template_name, :string
       remove_column table_name.to_sym, :schema, :jsonb
       remove_column (table_name.singularize + '_histories').to_sym, :template_name, :string
